moved new type to v2 mod
